+++ packages/docs/src/data/api/rules/max-length/arguments.ts
@@ -1,0 +1,18 @@
+import { Api } from '~/types';
+
+export const api: Api = {
+    maxLength: {
+		argument: [
+			{
+				name: 'max',
+				type: 'number',
+				description: 'La longueur maximum souhaitée'
+			},
+			{
+				name: 'errorMessages',
+				type: 'object',
+				description: 'Message d’erreurs'
+			}
+		]
+	}
+};