+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   - **PaginatedTable** Ajout de l'interface DataOptions dans l'API du composant ([#1401](hhttps://github.com/assurance-maladie-digital/design-system/pull/1401)) ([56f831f](https://github.com/assurance-maladie-digital/design-system/commit/56f831f91cb16e75bc6aab9fb98d22421d8c93bd))
   - **rules** Documentation de la règle de validation required ([#1412](https://github.com/assurance-maladie-digital/design-system/pull/1412)) ([4da4471](https://github.com/assurance-maladie-digital/design-system/commit/4da44715c5e7e8bda7d76d39ec97800efc686465))
 
+  - **MaxLength rule** Ajout de la règle de validation maxLength ([#1431](https://github.com/assurance-maladie-digital/design-system/pull/1431) ([d097925](https://github.com/assurance-maladie-digital/design-system/pull/1431/commits/d0979251a5a16394dc771289349006c45039b409))
 ### Interne
 
 - 🔒 **Sécurité**